+++ EstimoteSDKv5/EstimoteSDKv5.framework/Versions/A/Headers/EstimoteSDKv5.h
@@ -7,9 +7,11 @@
 //  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
 //
 //
-//  Version: 5.0.0-alpha.1
+//  Version: 5.0.0-alpha.2
 //  Copyright (c) 2017 Estimote. All rights reserved.
 
 #import "ESTCloudCredentials.h"
 #import "ESTProximityObserver.h"
 #import "ESTProximityRange.h"
+#import "ESTProximityZone.h"
+#import "ESTProximityDeviceAttachment.h"