+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 5.0.0-alpha.2 (October 6, 2017)
+
+Features: 
+
+- `ESTProximityZone` that allows filtering JSON attachments by key, value pairs. Instances of this class also enable registering callbacks for monitoring events.
+- Sending Estimote Monitoring analytics events that will be presented in Estimote Cloud's dashboard in upcoming release.
+
+Improvements: 
+
+- Optimized resources allocation when starting observation of multiple zones (calling `-[ESTProximityObserver startObservingZones:]`).
+
 ## 5.0.0-alpha.1 (September 15, 2017)
 
 Features: